+++ src/dreamstate/core/services/getReactContext.ts
@@ -1,17 +1,24 @@
 import { Context, createContext } from "react";
 
 import { CONTEXT_REACT_CONTEXTS_REGISTRY } from "@/dreamstate/core/internals";
-import { TAnyContextManagerConstructor } from "@/dreamstate/types";
+import { IContextManagerConstructor } from "@/dreamstate/types";
 
 /**
  * Get react context reference for provided ManagerClass.
  * Lazy internal creation only after first access attempt.
+ *
+ * @param ManagerClass - context manager constructor reference.
+ * @param defaultContext - default context value to be applied if manager is not provided.
+ * @returns {Context} react context instance with pre-defined default value.
  */
-export function getReactContext<T extends TAnyContextManagerConstructor>(ManagerClass: T): Context<T> {
+export function getReactContext<S, M extends IContextManagerConstructor<S>>(
+  ManagerClass: M,
+  defaultContext: Partial<S> | null = null
+): Context<S> {
   if (CONTEXT_REACT_CONTEXTS_REGISTRY.has(ManagerClass)) {
-    return CONTEXT_REACT_CONTEXTS_REGISTRY.get(ManagerClass) as Context<T>;
+    return CONTEXT_REACT_CONTEXTS_REGISTRY.get(ManagerClass) as Context<S>;
   } else {
-    const reactContext: Context<T> = createContext(null as any);
+    const reactContext: Context<S> = createContext(defaultContext as S);
 
     /**
      * Later providers and consumers in tree will be displayed as